Bugfix start in obstacle

+++ core/src/main/java/pfg/kraken/astar/AStarNode.java
@@ -29,8 +29,8 @@ public final class AStarNode implements Memorizable, Printable
 	private static final long serialVersionUID = -2120732124823178009L;
 	public Kinematic cinematique; // the cinematic state + the duration since the beginning of the search
 	public volatile long date = 0;
-	public int g_score; // distance du point de départ à ce point
-	public int f_score; // g_score + heuristique = meilleure distance qu'on
+	public double g_score; // distance du point de départ à ce point
+	public double f_score; // g_score + heuristique = meilleure distance qu'on
 							// peut espérer avec ce point
 	public AStarNode parent; // the parent of this node (used for reconstruction when a path is found)
